Change error handling during cue parsing
If an error is throw during cue parsing it's included in a `errors` array instead of stopping the entire parsing process.

The error array is returned and if it's not empty then `valid` property will be false. Errors throw outside cue parsing - like invalid signatures - still stop the parsing process as those aren't ignored during rendering.

Test cases were updated with this behavior.